requirements/base.txt
# SHA1:86025a6254cdaea6a94a2557c59e924eab7aa9b0
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
aiofiles==0.5.0           # via -r base.in
aiogrpc==1.7              # via -r base.in
grpcio==1.28.1            # via -r base.in, aiogrpc
protobuf==3.6.1           # via -r base.in
six==1.14.0               # via grpcio, protobuf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
